Move attachmentHelpers to helpers folder
