+++ packages/react-sdk-components/src/components/helpers/attachmentHelpers.js
